Move `libnetwork-setkey` from reexec to real cmd
This is called in a hot path (every container start) and as such we can
gain a huge performance boost by splitting it out into a separate
command so that it's not encumbered with initializing random packages
from the caller.
